src/requirements.txt
geoip2
kafka-python
certifi
wheel
pymongo
pytest
pylru
scrapy
dnspython
BeautifulSoup4
jsbeautifier>1.10.2
#pyod
psutil
#sqlite3
